Move state down to local components
